backend/requirements.txt
asgiref==3.3.2
Django==3.2
django-extra-fields==3.0.2
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.1
gunicorn==20.0.4
Pillow==9.5.0
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2020.1
requests==2.26.0
requests-oauthlib==1.3.1
sqlparse==0.3.1
